+++ Boolean.CSharp.Main/Accounts/AccountTypes.cs
@@ -1,4 +1,4 @@
-﻿namespace Boolean.CSharp.Main
+﻿namespace Boolean.CSharp.Main.Accounts
 {
     public enum AccountTypes
     {
