+++ Boolean.CSharp.Main/Customer.cs
@@ -1,6 +1,18 @@
-﻿namespace Boolean.CSharp.Main
+﻿
+namespace Boolean.CSharp.Main
 {
     public class Customer
     {
+        private string _name;
+
+        public Customer(string name)
+        {
+            _name = name;
+        }
+
+        public string GetName()
+        {
+            throw new NotImplementedException();
+        }
     }
 }
